+++ oid4vp-java/oid4vp-core/src/main/java/de/eecc/oid4vc/oid4vp/request/PresentationRequestDefinition.java
@@ -1,7 +1,9 @@
 package de.eecc.oid4vc.oid4vp.request;
 
+import com.fasterxml.jackson.databind.JsonNode;
 import de.eecc.oid4vc.oid4vp.ClientMetadata;
 import de.eecc.oid4vc.oid4vp.DcqlQuery;
+import de.eecc.oid4vc.oid4vp.PresentationClaims;
 
 /**
  * Credential-specific parts of an OID4VP authorization request.
@@ -14,4 +16,9 @@ public interface PresentationRequestDefinition {
     default ClientMetadata clientMetadata() {
         return ClientMetadata.presentationDefault();
     }
+
+    /**
+     * Extracts claims from a verified {@code vp_token} according to this definition's DCQL query.
+     */
+    PresentationClaims extractPresentationClaims(JsonNode vpTokenNode);
 }